requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
langchain==0.1.9
langchain-openai==0.0.5
langchain-pinecone==0.0.3
pinecone-client==3.0.3
openai==1.12.0
tiktoken==0.6.0
python-dotenv==1.0.0
